refactor(class-module): pagination response
