+++ src/api/class/class.controller.ts
@@ -64,29 +64,8 @@ export class ClassController {
   @ApiBearerAuth()
   @Get()
   @ResponseMessage('Get Classes successfully')
-  async getClasses(@Query() query: GetClassesQueryDto): Promise<OffsetPaginatedDto<ClassDTO>> {
-    const { classes, meta } = await this.classService.getClasses(query)
-
-    const mappedClasses = classes.map((c) =>
-      plainToInstance(ClassDTO, {
-        class_id: c.classId,
-        class_code: c.classCode,
-        class_name: c.className,
-        class_type: c.classType,
-        start_date: c.startDate,
-        end_date: c.endDate,
-        is_active: c.isActive,
-        created_at: c.createdAt,
-        updated_at: c.updatedAt
-      })
-    )
-
-    return new OffsetPaginatedDto<ClassDTO>({
-      statusCode: 200,
-      message: 'Get Classes successfully',
-      data: mappedClasses,
-      meta
-    })
+  async getClasses(@Query() query: GetClassesQueryDto) {
+    return await this.classService.getClasses(query)
   }
 
   @Get(':classId')
