+++ src/api/module/module.controller.ts
@@ -84,15 +84,8 @@ export class ModulesController {
   @ApiOperation({ summary: 'Get Modules', description: 'Get list of modules with pagination, search and filters' })
   @Get()
   @ResponseMessage('Get Modules successfully')
-  async getModules(@Query() query: GetModulesQueryDto): Promise<OffsetPaginatedDto<unknown>> {
-    const { modules, meta } = await this.modulesService.getModules(query)
-
-    return new OffsetPaginatedDto<unknown>({
-      statusCode: 200,
-      message: 'Get Modules successfully',
-      data: modules as unknown[],
-      meta
-    })
+  async getModules(@Query() query: GetModulesQueryDto) {
+    return await this.modulesService.getModules(query)
   }
 
   @Get(':moduleId')
