+++ src/api/module/module.service.ts
@@ -169,7 +169,7 @@ export class ModulesService {
 
     return {
       modules,
-      meta: metaDto
+      pagination: metaDto
     }
   }
